Moved EXR Half impl from math to new external folder
