+++ ext/openexr/openvdb/ext/Half.cc
